Update PSNR SSIM

+++ dataset/degradation.py
@@ -743,7 +743,7 @@ def add_jpg_compression(img, quality=90):
     """
     img = np.clip(img, 0, 1)
     encode_param = [int(cv2.IMWRITE_JPEG_QUALITY), quality]
-    _, encimg = cv2.imencode('.jpg', img * 255., encode_param)
+    _, encimg = cv2.imencode('.jpg', (img * 255.).astype(np.uint8), encode_param)
     img = np.float32(cv2.imdecode(encimg, 1)) / 255.
     return img
